Page Connect Project
This project implements the use of $_SESSION, $_SERVER, and $key variables to connect different PHP pages to one another if the link to that specific page is clicked.  Each page will also display the page the user previously visited, and a list of pages the user has visited during that current session.